pyproject.toml
[project]
name = "l46-mlsys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "faker>=40.11.0",
    "llama-cpp-python>=0.3.16",
    "psutil>=7.2.2",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "zeus>=0.15.0",
]

[dependency-groups]
cuda = [
    "nvidia-ml-py>=13.590.48",
    "pynvml>=13.0.1",
]
dev = [
    "pyrefly>=0.57.1",
    "ruff>=0.15.6",
    "ty>=0.0.23",
]
